# specs/design/api-contracts.schema.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OnboardX API",
    "version": "1.0.0",
    "description": "Customer onboarding and KYC platform (synthetic data). Integer-only numerics. See specs/design/api-contracts.md."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "local verification mode"
    }
  ],
...
  "paths": {
...
    "/api/v1/cases/{case_id}/documents": {
      "post": {
        "operationId": "uploadDocument",
        "summary": "Upload or re-upload a document",
        "tags": [
          "cases"
        ],
        "x-story": "E2-S1",
        "x-roles": [
          "prospect(owner)"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Classified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            }
          },
          "409": {
            "description": "CASE_LOCKED or INVALID_STATE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "FILE_TOO_LARGE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "UNSUPPORTED_MEDIA_TYPE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "UNKNOWN_CHECKLIST_ITEM or validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown case",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/DocumentUploadRequest"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listDocuments",
        "summary": "Current document per item",
        "tags": [
          "cases"
        ],
        "x-story": "E2-S1",
        "x-roles": [
          "prospect(owner)",
          "kyc-analyst",
          "compliance-officer",
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Documents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentList"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "include_superseded",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ]
      }
    },
    "/api/v1/cases/{case_id}/documents/{document_id}/reject": {
      "post": {
        "operationId": "rejectDocument",
        "summary": "Analyst rejects a document",
        "tags": [
          "cases"
        ],
        "x-story": "E2-S4",
        "x-roles": [
          "kyc-analyst"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectDocumentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "CASE_LOCKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Missing or unknown reason",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectDocumentRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cases/{case_id}/evidence": {
      "get": {
        "operationId": "getEvidence",
        "summary": "Screening, risk and documents for review",
        "tags": [
          "cases"
        ],
        "x-story": "E3-S5",
        "x-roles": [
          "kyc-analyst",
          "compliance-officer",
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Evidence",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evidence"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/review-queue": {
      "get": {
        "operationId": "reviewQueue",
        "summary": "MANUAL_REVIEW cases oldest first",
        "tags": [
          "review-queue"
        ],
        "x-story": "E4-S3",
        "x-roles": [
          "compliance-officer"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Queue",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewQueue"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/override": {
      "post": {
        "operationId": "overrideCase",
        "summary": "Audited override of MANUAL_REVIEW",
        "tags": [
          "cases"
        ],
        "x-story": "E4-S3",
        "x-roles": [
          "compliance-officer"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Overridden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverrideResponse"
                }
              }
            }
          },
          "409": {
            "description": "INVALID_STATE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Missing or unknown reason",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/reclassify": {
      "post": {
        "operationId": "reclassifyCase",
        "summary": "Officer risk-band reassignment",
        "tags": [
          "cases"
        ],
        "x-story": "E4-S3",
        "x-roles": [
          "compliance-officer"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "New assessment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskAssessment"
                }
              }
            }
          },
          "409": {
            "description": "INVALID_STATE or CASE_LOCKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Missing reason",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReclassifyRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Product": {
        "type": "string",
        "enum": [
          "Savings",
          "Current",
          "NRE"
        ]
      },
      "CaseState": {
        "type": "string",
        "enum": [
          "INITIATED",
          "DOCS_SUBMITTED",
          "SCREENED",
          "CLASSIFIED",
          "APPROVED",
          "REJECTED",
          "MANUAL_REVIEW"
        ]
      },
...
      "ChecklistItemCode": {
        "type": "string",
        "enum": [
          "ID_PROOF",
          "ADDRESS_PROOF",
          "PHOTOGRAPH",
          "BUSINESS_PROOF",
          "OVERSEAS_ADDRESS_PROOF"
        ]
      },
      "DocClass": {
        "type": "string",
        "enum": [
          "PAN",
          "AADHAAR",
          "PASSPORT",
          "UTILITY_BILL",
          "PHOTOGRAPH",
          "GST_CERTIFICATE",
          "VISA",
          "UNRECOGNISED"
        ]
      },
...
      "DocRejectReasonCode": {
        "type": "string",
        "enum": [
          "DOC_ILLEGIBLE",
          "DOC_EXPIRED",
          "DOC_NAME_MISMATCH",
          "DOC_WRONG_TYPE",
          "DOC_OTHER"
        ]
      },
      "RiskBand": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH"
        ]
      },
      "ReviewReasonCode": {
        "type": "string",
        "enum": [
          "AML_HIT",
          "PEP_HIT",
          "DOC_UNRECOGNISED",
          "RISK_MEDIUM",
          "RISK_HIGH"
        ]
      },
      "OverrideReasonCode": {
        "type": "string",
        "enum": [
          "FALSE_POSITIVE_CLEARED",
          "RISK_ACCEPTED",
          "DOCS_CONFIRMED",
          "CONFIRMED_WATCHLIST_MATCH",
          "DOCS_INSUFFICIENT",
          "RISK_TOO_HIGH",
          "POLICY_OTHER"
        ]
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "UNAUTHENTICATED",
          "FORBIDDEN",
          "NOT_FOUND",
          "INVALID_STATE",
          "CASE_LOCKED",
          "PROFILE_LOCKED",
          "RULESET_IMMUTABLE",
          "FILE_TOO_LARGE",
          "UNSUPPORTED_MEDIA_TYPE",
          "VALIDATION_ERROR",
          "UNKNOWN_CHECKLIST_ITEM",
          "MISSING_DOCUMENTS",
          "MISSING_PROFILE",
          "MISSING_PROFILE_FIELD",
          "RULESET_INVALID",
          "UNKNOWN_REASON_CODE",
          "DRAFT_EXISTS",
          "ALREADY_DEACTIVATED"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "$ref": "#/components/schemas/ErrorCode"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            },
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "details"
            ]
          }
        },
        "additionalProperties": false,
        "required": [
          "error"
        ]
      },
...
      "DocumentView": {
        "type": "object",
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid"
          },
          "checklist_item": {
            "$ref": "#/components/schemas/ChecklistItemCode"
          },
          "version": {
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/DocStatus"
          },
          "doc_class": {
            "$ref": "#/components/schemas/DocClass"
          },
          "reason_code": {
            "type": "string",
            "nullable": true
          },
          "confidence_bp": {
            "type": "integer"
          },
          "rule_version": {
            "type": "integer"
          },
          "superseded": {
            "type": "boolean"
          },
          "size_bytes": {
            "type": "integer"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "required": [
          "document_id",
          "checklist_item",
          "version",
          "status",
          "doc_class",
          "reason_code",
          "confidence_bp",
          "rule_version",
          "superseded",
          "size_bytes",
          "sha256",
          "uploaded_at"
        ]
      },
      "DocumentList": {
        "type": "object",
        "properties": {
          "case_id": {
            "type": "string",
            "format": "uuid"
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentView"
            }
          }
        },
        "additionalProperties": false,
        "required": [
          "case_id",
          "documents"
        ]
      },
      "DocumentUploadRequest": {
        "type": "object",
        "properties": {
          "checklist_item": {
            "$ref": "#/components/schemas/ChecklistItemCode"
          },
          "file": {
            "type": "string",
            "format": "binary"
          }
        },
        "additionalProperties": false,
        "required": [
          "checklist_item",
          "file"
        ]
      },
      "DocumentUploadResponse": {
        "type": "object",
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid"
          },
          "checklist_item": {
            "$ref": "#/components/schemas/ChecklistItemCode"
          },
          "version": {
            "type": "integer"
          },
          "doc_class": {
            "$ref": "#/components/schemas/DocClass"
          },
          "status": {
            "type": "string",
            "enum": [
              "VERIFIED",
              "FLAGGED"
            ]
          },
          "reason_code": {
            "type": "string",
            "nullable": true
          },
          "confidence_bp": {
            "type": "integer"
          },
          "rule_version": {
            "type": "integer"
          }
        },
        "additionalProperties": false,
        "required": [
          "document_id",
          "checklist_item",
          "version",
          "doc_class",
          "status",
          "reason_code",
          "confidence_bp",
          "rule_version"
        ]
      },
      "RejectDocumentRequest": {
        "type": "object",
        "properties": {
          "reason_code": {
            "$ref": "#/components/schemas/DocRejectReasonCode"
          },
          "comment": {
            "type": "string",
            "maxLength": 500
          }
        },
        "additionalProperties": false,
        "required": [
          "reason_code"
        ]
      },
      "RejectDocumentResponse": {
        "type": "object",
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "REJECTED"
            ]
          },
          "reason_code": {
            "$ref": "#/components/schemas/DocRejectReasonCode"
          }
        },
        "additionalProperties": false,
        "required": [
          "document_id",
          "status",
          "reason_code"
        ]
      },
...
      "ScreeningResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "case_id": {
            "type": "string",
            "format": "uuid"
          },
          "hits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScreeningHit"
            }
          },
          "requires_manual_review": {
            "type": "boolean"
          },
          "watchlist_version": {
            "type": "integer"
          },
          "screened_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "required": [
          "id",
          "case_id",
          "hits",
          "requires_manual_review",
          "watchlist_version",
          "screened_at"
        ]
      },
...
      "RiskAssessment": {
        "type": "object",
        "properties": {
          "assessment_id": {
            "type": "string",
            "format": "uuid"
          },
          "case_id": {
            "type": "string",
            "format": "uuid"
          },
          "score": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "band": {
            "$ref": "#/components/schemas/RiskBand"
          },
          "rule_version": {
            "type": "integer"
          },
          "source": {
            "type": "string",
            "enum": [
              "RULE_ENGINE",
              "OFFICER_RECLASSIFY"
            ]
          },
          "breakdown": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RiskBreakdownItem"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "required": [
          "assessment_id",
          "case_id",
          "score",
          "band",
          "rule_version",
          "source",
          "breakdown",
          "created_at"
        ]
      },
...
      "Decision": {
        "type": "object",
        "properties": {
          "decision_id": {
            "type": "string",
            "format": "uuid"
          },
          "case_id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "AUTO"
            ]
          },
          "outcome": {
            "type": "string",
            "enum": [
              "APPROVED",
              "MANUAL_REVIEW"
            ]
          },
          "reason_code": {
            "type": "string",
            "description": "AUTO_APPROVED or ReviewReasonCode"
          },
          "rule_version": {
            "type": "integer"
          },
          "actor": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "required": [
          "decision_id",
          "case_id",
          "type",
          "outcome",
          "reason_code",
          "rule_version",
          "actor",
          "created_at"
        ]
      },
...
      "Evidence": {
        "type": "object",
        "properties": {
          "case_id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "$ref": "#/components/schemas/CaseState"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentView"
            }
          },
          "screening": {
            "$ref": "#/components/schemas/ScreeningResult",
            "nullable": true
          },
          "risk_assessment": {
            "$ref": "#/components/schemas/RiskAssessment",
            "nullable": true
          },
          "decision": {
            "$ref": "#/components/schemas/Decision",
            "nullable": true
          },
          "override": {
            "$ref": "#/components/schemas/OverrideView",
            "nullable": true
          },
          "review_reason_code": {
            "$ref": "#/components/schemas/ReviewReasonCode",
            "nullable": true
          },
          "account_number_masked": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "required": [
          "case_id",
          "state",
          "product",
          "documents",
          "screening",
          "risk_assessment",
          "decision",
          "override",
          "review_reason_code",
          "account_number_masked"
        ]
      },
...
      "ReviewQueue": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewQueueItem"
            }
          },
          "total": {
            "type": "integer"
          }
        },
        "additionalProperties": false,
        "required": [
          "items",
          "total"
        ]
      },
      "OverrideRequest": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "APPROVE",
              "REJECT"
            ]
          },
          "reason_code": {
            "$ref": "#/components/schemas/OverrideReasonCode"
          },
          "comment": {
            "type": "string",
            "maxLength": 500
          }
        },
        "additionalProperties": false,
        "required": [
          "decision",
          "reason_code"
        ]
      },
      "OverrideResponse": {
        "type": "object",
        "properties": {
          "case_id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "enum": [
              "APPROVED",
              "REJECTED"
            ]
          },
          "override": {
            "$ref": "#/components/schemas/OverrideView"
          },
          "account_number": {
            "type": "string",
            "pattern": "^(SAV|CUR|NRE)[0-9]{12}$",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "required": [
          "case_id",
          "state",
          "override",
          "account_number"
        ]
      },
      "ReclassifyRequest": {
        "type": "object",
        "properties": {
          "band": {
            "$ref": "#/components/schemas/RiskBand"
          },
          "reason_code": {
            "$ref": "#/components/schemas/ReclassifyReasonCode"
          },
          "comment": {
            "type": "string",
            "maxLength": 500
          }
        },
        "additionalProperties": false,
        "required": [
          "band",
          "reason_code"
        ]
      },
...
    }
  }
}